+++ packages/stream_video/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 ✅ Added
+* Added `setMirrorVideo` method to `Call` class to control video mirroring for participants.
 * Added `call.partialState` for more specific and efficient state updates.
 
 ## 0.9.6